cleanup dashbaord container
